pyproject.toml
[project]
name = "editorbar"
version = "0.1.0"
description = "Turns the default Outliner and Properties editors in Blender workspaces into a sidebar that you can quickly collapse and expand"
readme = 'README.md'
requires-python = '>=3.12'
dependencies = []

[dependency-groups]
dev = [
    "fake-bpy-module-latest>=20251003",
]
